move insecureHTTPParser tests to a dedicated file
node-http.test.ts carries an unrelated non-hermetic proxy test that
fails in sandboxed environments; split the new coverage into its own
topic-specific file like the rest of this directory.
